refactor(tests): upgrade test suite
