+++ app/upgrades/v7/integration/network.go
@@ -1,4 +1,4 @@
-package tests
+package integration
 
 import (
 	upgradetypes "cosmossdk.io/x/upgrade/types"
@@ -7,6 +7,7 @@ import (
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
+	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	erc20types "github.com/evmos/evmos/v20/x/erc20/types"
 	evmtypes "github.com/evmos/evmos/v20/x/evm/types"
@@ -64,6 +65,10 @@ func (n *UpgradeTestNetwork) GetAuthzClient() authz.QueryClient {
 
 func (n *UpgradeTestNetwork) GetStakingClient() stakingtypes.QueryClient {
 	return exrpcommon.GetStakingClient(n)
+}
+
+func (n *UpgradeTestNetwork) GetSlashingClient() slashingtypes.QueryClient {
+	return exrpcommon.GetSlashingClient(n)
 }
 
 func (n *UpgradeTestNetwork) GetDistrClient() distrtypes.QueryClient {
